everything except autofill

+++ webapp/src/app/components/header/header.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-header',
@@ -16,7 +17,19 @@ export class Header {
   }
 
   logout() {
-    localStorage.clear();
-    this.router.navigate(['/login']);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You will be logged out",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#5daaf3ff',
+      cancelButtonColor: 'rgba(221, 71, 51, 1)',
+      confirmButtonText: 'Logout'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        localStorage.clear();      // remove token
+        this.router.navigate(['/login']);
+      }
+    });
   }
 }
